Script that will restore the history
This script will restore the history that was lost when the beta version was copied.

  +-----------------+
  |  Current state  |
  +-----------------+                +--------------------+
  ---------------------------------> | Release Repository |
                                     +--------------------+
                                     +--------------------+
                  -----------------> |  Beta Repository   |
                                     +--------------------+

 +-----------------------+
 | After running script  |
 +-----------------------+           +--------------------+
  ----------------+----------------> | Release Repository |
                  |                  +--------------------+
                  |
                  |                  +--------------------+
                  +----------------> |  Beta Repository   |
                                     +--------------------+

The script will restore the history, without actually merging any changes
that have happened on the release branch since the beta repo was created.

Because this does re-write the history, you should know what you are doing
before you run it. This script contains a protection that will prevent it
from making any changes.